+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changelog
 -----------
 
 - Added syslog-ng support (#2)
+- Added sudo arg to utils.popen_wrapper() - code consolidation
 - Fixed splitlines list[bytes] decode on Python 3
+- Logging output tweaks
+- Fixed redundant TCPDump.check_packet_print() in nfsinkhole-setup.py
 
 0.1.0 (2016-08-29)
 ------------------